+++ kernel/include/kernel/memory/kheap.h
@@ -5,7 +5,7 @@
 extern "C" {
 #endif
 
-    void* kmalloc(uint32_t size);
+    void* kmalloc(uint64_t size);
     void kfree(void* ptr);
 
 #ifdef __cplusplus
